requirements/base.txt
fastapi==0.110.1
uvicorn==0.29.0
asyncpg==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
bcrypt==4.1.2
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
celery==5.3.6
jinja2==3.1.3
email-validator>=2.0.0
redis==5.0.3
redis[hiredis]
pyTelegramBotAPI==4.17.0
grpcio-tools==1.62.1
minio==7.2.5